Clean out pre-0.4 cruft
