+++ test/deprecated.jl
@@ -1,0 +1,45 @@
+v = 1
+d = Dict{Int,Int}()
+d[1] = 1
+@test Compat.@Dict(1 => 1) == d
+@test Compat.@Dict(1 => v) == d
+
+@test typeof(@compat(Dict(1 => 1))) == Dict{Int,Int}
+@test @compat(Dict(1 => 1)) == d
+@test @compat(Dict(1 => v)) == d
+
+ad = Dict{Any,Any}()
+ad[1] = 1
+@test Compat.@AnyDict(1 => 1) == ad
+@test Compat.@AnyDict(1 => v) == ad
+
+@test typeof(@compat(Dict{Any,Any}(1 => 1))) == Dict{Any,Any}
+@test @compat(Dict{Any,Any}(1 => 1)) == ad
+@test @compat(Dict{Any,Any}(1 => v)) == ad
+
+td = Dict{Int,Float64}()
+td[1] = 1.0
+
+@test typeof(@compat(Dict{Int,Float64}(1 => 1))) == Dict{Int,Float64}
+@test @compat(Dict{Int,Float64}(1 => 1)) == td
+@test @compat(Dict{Int,Float64}(1 => v)) == td
+
+@test @compat(Dict()) == Dict()
+@test @compat(Dict{Any,Any}()) == Dict{Any,Any}()
+@test @compat(Dict([(1, 1)])) == d
+@test @compat(Dict(:Void => :Nothing)) == Dict([(:Void, :Nothing)])
+
+d2 = Dict{Symbol,Dict{Symbol,Int}}()
+d2[:a] = Dict{Symbol,Int}()
+d2[:a][:b] = 1
+@test @compat(Dict(:a => Dict(:b => 1))) == d2
+
+d = Dict(zip([1, 2], [3, 4]))
+@test d == @compat Dict(1=>3, 2=>4)
+
+@compat function f()
+    a = :a
+    b = Dict(:b => 1)
+    Dict(a => b)
+end
+@test f() == d2
